+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.23.0 - 2023-10-26
+
+#### Enhancements
+- Added `map_symbols_csv` function to the `databento` module for using `symbology.json` files to map a symbol column onto a CSV file
+- Added `map_symbols_json` function to the `databento` module for using `symbology.json` files to add a symbol key to a file of JSON records
+- Added new publisher values in preparation for IFEU.IMPACT and NDEX.IMPACT datasets
+
+#### Bug fixes
+- Fixed issue where a large unreadable symbol subscription message could be sent
+- Fixed an issue where `DBNStore.to_df` with `pretty_ts=True` was very slow
+
 ## 0.22.1 - 2023-10-24
 
 #### Bug fixes